requirements.txt
-e .
aws-cdk.aws-ec2 == 1.57.0
aws-cdk.aws-s3-assets == 1.57.0
Jinja2 == 2.11.2